Pass correct trigger reference in dispatch
When writing a programmable sensor (i.e. making use of the *_trigger
methods) it seems that it is not the name of the trigger that should be
used, but the trigger reference. This makes sense since a sensor could
have multiple sensors of the same type associated with it, and using the
name would make it impossible to tell them apart. However, I haven't
found this documented anywhere.