Check for copycomplete on remote hosts
This makes it possible to check for copycomplete files on remote
hosts. In addition, it saves the run directories from the dispatched
events in the key-value datastore in order to not dispatch duplicates.

The sensor will be reworked to detect run directories and their state as
opposed to only copycomplete files. In addition, it should only keep the
run directories that currently exist within the watched directories in
order to not become swamped.